+++ kbfscrypto/errors.go
@@ -1,0 +1,30 @@
+// Copyright 2016 Keybase Inc. All rights reserved.
+// Use of this source code is governed by a BSD
+// license that can be found in the LICENSE file.
+
+package kbfscrypto
+
+import (
+	"fmt"
+
+	"github.com/keybase/client/go/protocol/keybase1"
+)
+
+// InvalidKIDError is returned whenever an invalid KID is detected.
+type InvalidKIDError struct {
+	kid keybase1.KID
+}
+
+func (e InvalidKIDError) Error() string {
+	return fmt.Sprintf("Invalid KID %s", e.kid)
+}
+
+// InvalidByte32DataError is returned whenever invalid data for a
+// 32-byte type is detected.
+type InvalidByte32DataError struct {
+	data []byte
+}
+
+func (e InvalidByte32DataError) Error() string {
+	return fmt.Sprintf("Invalid byte32 data %v", e.data)
+}
